+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 * [#1786](https://github.com/shlinkio/shlink/issues/1786) Run API tests with RoadRunner by default.
 * [#2008](https://github.com/shlinkio/shlink/issues/2008) Update to Doctrine ORM 3.0.
 * [#2010](https://github.com/shlinkio/shlink/issues/2010) Update to Symfony 7.0 components.
+* [#2016](https://github.com/shlinkio/shlink/issues/2016) Simplify and improve how code coverage is generated in API and CLI tests".
 
 ### Deprecated
 * *Nothing*